Fix update category dan menambahkan sweetalert2 pada menuController

+++ app/Http/Controllers/MenuController.php
@@ -2,12 +2,12 @@
 
 namespace App\Http\Controllers;
 
-use App\Http\Requests\MenuRequest;
 use App\Http\Requests\MenuUpdateRequest;
 use App\Models\Menu_Kue;
 use App\Models\Menu_Kue_Kering;
 use App\Models\Menu_Nasi;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\File;
 use Illuminate\Support\Str;
 
 class MenuController extends Controller
@@ -19,9 +19,13 @@ class MenuController extends Controller
     {
         //
         $menuKueLoyang = Menu_Kue::all();
+        $jumlahKL = Menu_Kue::count();
         $menuKueKering = Menu_Kue_Kering::all();
+        $jumlahKK = Menu_Kue_Kering::count();
         $menuNasi = Menu_Nasi::all();
-        return view('admin.index', compact('menuKueLoyang', 'menuKueKering', 'menuNasi'));
+        $jumlahNasi = Menu_Nasi::count();
+
+        return view('admin.index', compact('menuKueLoyang', 'menuKueKering', 'menuNasi', 'jumlahKL', 'jumlahKK', 'jumlahNasi'));
     }
 
     /**
@@ -36,16 +40,33 @@ class MenuController extends Controller
     /**
      * Store a newly created resource in storage.
      */
-    public function store(MenuRequest $request)
+    public function store(Request $request)
     {
         //
         $kategori = $request->input('kategori');
 
-        $validated = $request->validated();
-
         switch ($kategori) {
             case 'kue':
                 // Ambil nama asli file, kl ga pake ini namanya di generate acak sm laravel
+                $validate = $request->validate([
+                    'name' => ['required', 'min:3', 'max:100', 'unique:menu_kue,name'],
+                    'harga_normal' => ['required', 'numeric', 'min:400'],
+                    'deskripsi' => ['required'],
+                    'image' => ['required', 'file', 'mimes:png,jpg,jpeg,svg', 'max:8092'],
+                ], [
+                    'name.required' => 'Username wajib diisi',
+                    'name.min' => 'Minimal 3 huruf',
+                    'name.max' => 'Maksimal 100 huruf',
+                    'name.unique' => 'Nama menu sudah pernah terdaftar!',
+                    'harga_normal.required' => 'Harga wajib diisi',
+                    'harga_normal.min' => 'Angka yang diinput terlalu kecil! (Minimal Rp.400)',
+                    'harga_normal.numeric' => 'Masukan angka saja',
+                    'deskripsi.required' => 'Deskripsi wajib diisi',
+                    'image.required' => "Mohon upload gambar melalui tombol 'choose file'",
+                    'image.mimes' => 'Ekstensi yang diperbolehkan hanya:png,jpg,jpeg,svg',
+                    'image.max' => 'Ukuran maksimum gambar adalah 8mb (8092kb)!',
+                ]);
+
                 $fileName = $request->image->getClientOriginalname();
 
                 $data = new Menu_Kue;
@@ -58,9 +79,28 @@ class MenuController extends Controller
 
                 // Taro hard copy file ke dalam folder thumbnail yg ada di public
                 $request->image->storeAs('menu_kue_loyang', $fileName);
-                return redirect(route('menu.create'))->with('flash_msg', 'Data berhasil ditambah!');
+                return redirect(route('menu.create'))->withSuccess('Menu berhasil ditambah!');
 
             case 'kue_kering':
+                $validate = $request->validate([
+                    'name' => ['required', 'min:3', 'max:100', 'unique:menu_kue_kering,name'],
+                    'harga_normal' => ['required', 'numeric', 'min:400'],
+                    'deskripsi' => ['required'],
+                    'image' => ['required', 'file', 'mimes:png,jpg,jpeg,svg', 'max:8092'],
+                ], [
+                    'name.required' => 'Username wajib diisi',
+                    'name.min' => 'Minimal 3 huruf',
+                    'name.max' => 'Maksimal 100 huruf',
+                    'name.unique' => 'Nama menu sudah pernah terdaftar!',
+                    'harga_normal.required' => 'Harga wajib diisi',
+                    'harga_normal.min' => 'Angka yang diinput terlalu kecil! (Minimal Rp.400)',
+                    'harga_normal.numeric' => 'Masukan angka saja',
+                    'deskripsi.required' => 'Deskripsi wajib diisi',
+                    'image.required' => "Mohon upload gambar melalui tombol 'choose file'",
+                    'image.mimes' => 'Ekstensi yang diperbolehkan hanya:png,jpg,jpeg,svg',
+                    'image.max' => 'Ukuran maksimum gambar adalah 8mb (8092kb)!',
+                ]);
+
                 $fileName = $request->image->getClientOriginalname();
 
                 $data = new Menu_Kue_Kering;
@@ -73,8 +113,27 @@ class MenuController extends Controller
 
                 // Taro hard copy file ke dalam folder thumbnail yg ada di public
                 $request->image->storeAs('menu_kue_kering', $fileName);
-                return redirect(route('menu.create'))->with('flash_msg', 'Data berhasil ditambah!');
+                return redirect(route('menu.create'))->withSuccess('Menu berhasil ditambah!');
+
             case 'nasi':
+                $validate = $request->validate([
+                    'name' => ['required', 'min:3', 'max:100', 'unique:menu_nasi,name'],
+                    'harga_normal' => ['required', 'numeric', 'min:400'],
+                    'deskripsi' => ['required'],
+                    'image' => ['required', 'file', 'mimes:png,jpg,jpeg,svg', 'max:8092'],
+                ], [
+                    'name.required' => 'Username wajib diisi',
+                    'name.min' => 'Minimal 3 huruf',
+                    'name.max' => 'Maksimal 100 huruf',
+                    'name.unique' => 'Nama menu sudah pernah terdaftar!',
+                    'harga_normal.required' => 'Harga wajib diisi',
+                    'harga_normal.min' => 'Angka yang diinput terlalu kecil! (Minimal Rp.400)',
+                    'harga_normal.numeric' => 'Masukan angka saja',
+                    'deskripsi.required' => 'Deskripsi wajib diisi',
+                    'image.required' => "Mohon upload gambar melalui tombol 'choose file'",
+                    'image.mimes' => 'Ekstensi yang diperbolehkan hanya:png,jpg,jpeg,svg',
+                    'image.max' => 'Ukuran maksimum gambar adalah 8mb (8092kb)!',
+                ]);
                 $fileName = $request->image->getClientOriginalname();
 
                 $data = new Menu_Nasi;
@@ -87,9 +146,10 @@ class MenuController extends Controller
 
                 // Taro hard copy file ke dalam folder thumbnail yg ada di public
                 $request->image->storeAs('menu_nasi', $fileName);
-                return redirect(route('menu.create'))->with('flash_msg', 'Data berhasil ditambah!');
+
+                return redirect(route('menu.create'))->withSuccess('Menu berhasil ditambah!');
             default:
-                return redirect(route('menu.create'))->with('flash_msg_error', 'Data gagal ditambah!');
+                return redirect(route('menu.create'))->withErrors("Menu gagal ditambah :(");
         }
     }
 
@@ -126,7 +186,35 @@ class MenuController extends Controller
 
         if (!$menu) {
             // Handle the case when the record with the given slug is not found
-            return redirect()->back()->with('flash_msg_error', 'Menu tidak ditemukan!');
+            return redirect()->back()->withErrors("Menu tidak ditemukan, silakan coba lagi... :(");
+        }
+
+        // Check if kategori not kue loyang
+        if ($request->kategori == "kue_kering") {
+            // Create the new record in Menu_Kue Table
+            $baru = Menu_Kue_Kering::create([$menu->toArray()]);
+
+            if ($baru) {
+                // Move the old file
+                File::move(public_path('storage/menu_kue_loyang/' . $menu->image), public_path('storage/menu_kue_kering/' . $menu->image));
+                // Delete the old record file
+                $menu->delete();
+                return redirect(route('menu.showKueKering', ['slug' => $slug]))->withSuccess('Menu berhasil diupdate!');
+            } else {
+                return redirect()->back()->withErrors("Menu gagal diupdate :(");
+            }
+
+        } else if ($request->kategori == "nasi") {
+            $baru = Menu_Nasi::create($menu->toArray());
+            if ($baru) {
+                // Move the old file
+                File::move(public_path('storage/menu_kue_loyang/' . $menu->image), public_path('storage/menu_nasi/' . $menu->image));
+                // Delete the old record file
+                $menu->delete();
+                return redirect(route('menu.showMenuNasi', ['slug' => $slug]))->withSuccess('Menu berhasil diupdate!');
+            } else {
+                return redirect()->back()->withErrors("Menu gagal diupdate :(");
+            }
         }
 
         // Update only the fields that are provided by the user
@@ -148,17 +236,47 @@ class MenuController extends Controller
 
         $menu->save();
 
-        return redirect(route('menu.showKueLoyang', ['slug' => $slug]))->with('flash_msg', 'Data berhasil diupdate!');
+        return redirect(route('menu.showKueLoyang', ['slug' => $slug]))->withSuccess('Menu berhasil diupdate!');
     }
-    public function updateKueKering(Request $request, string $slug)
+    public function updateKueKering(MenuUpdateRequest $request, string $slug)
     {
+
         // Validasi data
         $validated = $request->validated();
         $menu = Menu_Kue_Kering::where('slug', $slug)->first();
 
         if (!$menu) {
             // Handle the case when the record with the given slug is not found
-            return redirect()->back()->with('flash_msg_error', 'Menu tidak ditemukan!');
+            return redirect()->back()->withErrors("Menu gagal diupdate :(");
+        }
+
+        // Check if kategori not kue kering
+        if ($request->kategori == "kue") {
+            // Create the new record in Menu_Kue Table
+            $baru = Menu_Kue::create($menu->toArray());
+
+            // Kalau pembuatan data di tabel baru berhasil
+            if ($baru) {
+                // Move the old file
+                File::move(public_path('storage/menu_kue_kering/' . $menu->image), public_path('storage/menu_kue_loyang/' . $menu->image));
+                // Delete the old record file
+                $menu->delete();
+                return redirect(route('menu.showKueLoyang', ['slug' => $slug]))->withSuccess('Menu berhasil diupdate!');
+            } else {
+                return redirect()->back()->withErrors("Menu gagal diupdate :(");
+            }
+
+        } else if ($request->kategori == "nasi") {
+            $baru = Menu_Nasi::create($menu->toArray());
+            if ($baru) {
+                // Move the old file
+                File::move(public_path('storage/menu_kue_kering/' . $menu->image), public_path('storage/menu_nasi/' . $menu->image));
+                // Delete the old record file
+                $menu->delete();
+                return redirect(route('menu.showMenuNasi', ['slug' => $slug]))->withSuccess('Menu berhasil diupdate!');
+            } else {
+                return redirect()->back()->withErrors("Menu gagal diupdate :(");
+            }
         }
 
         // Update only the fields that are provided by the user
@@ -166,6 +284,7 @@ class MenuController extends Controller
             'name' => $request->input('name', $menu->name),
             'harga_normal' => $request->input('harga_normal', $menu->harga_normal),
             'deskripsi' => $request->input('deskripsi', $menu->deskripsi),
+
         ]);
 
         // Handle the image update
@@ -180,9 +299,9 @@ class MenuController extends Controller
 
         $menu->save();
 
-        return redirect(route('menu.showKueKering', ['slug' => $slug]))->with('flash_msg', 'Data berhasil diupdate!');
+        return redirect(route('menu.showKueKering', ['slug' => $slug]))->withSuccess("Data berhasil diupdate!");
     }
-    public function updateMenuNasi(Request $request, string $slug)
+    public function updateMenuNasi(MenuUpdateRequest $request, string $slug)
     {
         //
         // Validasi data
@@ -191,7 +310,36 @@ class MenuController extends Controller
 
         if (!$menu) {
             // Handle the case when the record with the given slug is not found
-            return redirect()->back()->with('flash_msg_error', 'Menu tidak ditemukan!');
+            return redirect()->back()->withErrors("Menu gagal diupdate :(");
+        }
+
+        // Check if kategori not nasi
+        if ($request->kategori == "kue") {
+            // Create the new record in Menu_Kue Table
+            $baru = Menu_Kue::create($menu->toArray());
+
+            // Kalau pembuatan data di tabel baru berhasil
+            if ($baru) {
+                // Move the old file
+                File::move(public_path('storage/menu_nasi/' . $menu->image), public_path('storage/menu_kue_loyang/' . $menu->image));
+                // Delete the old record file
+                $menu->delete();
+                return redirect(route('menu.showKueLoyang', ['slug' => $slug]))->withSuccess("Data berhasil diupdate!");
+            } else {
+                return redirect()->back()->withErrors("Menu gagal diupdate :(");
+            }
+
+        } else if ($request->kategori == "kue_kering") {
+            $baru = Menu_Kue_Kering::create($menu->toArray());
+            if ($baru) {
+                // Move the old file
+                File::move(public_path('storage/menu_nasi/' . $menu->image), public_path('storage/menu_kue_kering/' . $menu->image));
+                // Delete the old record file
+                $menu->delete();
+                return redirect(route('menu.showKueKering', ["slug" => $slug]))->withSuccess("Data berhasil diupdate!");
+            } else {
+                return redirect()->back()->withErrors("Menu gagal diupdate :(");
+            }
         }
 
         // Update only the fields that are provided by the user
@@ -199,6 +347,7 @@ class MenuController extends Controller
             'name' => $request->input('name', $menu->name),
             'harga_normal' => $request->input('harga_normal', $menu->harga_normal),
             'deskripsi' => $request->input('deskripsi', $menu->deskripsi),
+
         ]);
 
         // Handle the image update
@@ -212,8 +361,7 @@ class MenuController extends Controller
         }
 
         $menu->save();
-
-        return redirect(route('menu.showKueLoyang', ['slug' => $slug]))->with('flash_msg', 'Data berhasil diupdate!');
+        return redirect(route('menu.showMenuNasi', ['slug' => $slug]))->withSuccess("Data berhasil diupdate!");
     }
 
     /**
@@ -222,30 +370,41 @@ class MenuController extends Controller
     public function hapusMenuNasi($slug)
     {
         //
-        $data = Menu_Nasi::find($slug);
-        unlink(public_path('storage/menu_nasi/' . $data->image));
+        $data = Menu_Nasi::where('slug', $slug)->first();
 
-        $data->delete();
+        if ($data != null) {
+            unlink(public_path('storage/menu_nasi/' . $data->image));
+            $data->delete();
+            return redirect()->route('adminPage');
+        } else {
+            return redirect()->route('adminPage');
+        }
 
-        return redirect()->route('adminPage');
     }
 
     public function hapusMenuKueLoyang($slug)
     {
-        $data = Menu_Kue::find($slug);
-        unlink(public_path('storage/menu_kue_loyang/' . $data->image));
+        $data = Menu_Kue::where('slug', $slug)->first();
 
-        $data->delete();
-
-        return redirect()->route('adminPage');
+        if ($data != null) {
+            unlink(public_path('storage/menu_kue_loyang/' . $data->image));
+            $data->delete();
+            return redirect()->route('adminPage');
+        } else {
+            return redirect()->route('adminPage');
+        }
     }
     public function hapusMenuKueKering($slug)
     {
-        $data = Menu_Kue_Kering::find($slug);
-        unlink(public_path('storage/menu_kue_kering/' . $data->image));
+        $data = Menu_Kue_Kering::where('slug', $slug)->first();
 
-        $data->delete();
+        if ($data != null) {
+            unlink(public_path('storage/menu_kue_kering/' . $data->image));
 
-        return redirect()->route('adminPage');
+            $data->delete();
+            return redirect()->route('adminPage');
+        } else {
+            return redirect()->route('adminPage');
+        }
     }
 }
